Implemented Infusion RS232 Recorder Infusion data is now shown in table view Fixed regex for BIS and NIRS

+++ application/fxapp/src/main/java/de/cassisi/hearth/ui/event/AddInfusionDataEvent.java
@@ -23,6 +23,6 @@ public final class AddInfusionDataEvent {
     @AllArgsConstructor
     public static class PerfusionData {
         public final String name;
-        public final int rate;
+        public final double rate;
     }
 }
